# GatorFinder/Backend/docs/swagger.yaml
definitions:
  models.Event:
    properties:
      datePosted:
        type: string
      description:
        type: string
      endDate:
        type: string
      endTime:
        type: string
      id:
        type: string
      image:
        type: string
      name:
        type: string
      startDate:
        type: string
      startTime:
        type: string
      username:
        type: string
    type: object
  models.User:
    properties:
      email:
        type: string
      id:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
  /events/add:
    post:
      consumes:
      - application/json
      description: Adds a new event to the system
      parameters:
      - description: Event Data
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/models.Event'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Event'
      summary: Add a new event
      tags:
      - Events
  /events/delete:
    delete:
      consumes:
      - application/json
      description: Deletes an event from the system by ID
      parameters:
      - description: Event ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Event deleted successfully
          schema:
            type: string
      summary: Delete an event
      tags:
      - Events
  /events/get:
    get:
      consumes:
      - application/json
      description: Retrieves event details from the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Event'
            type: array
      summary: Get event details
      tags:
      - Events
  /events/upload:
    post:
      consumes:
      - image/png
      description: Uploads a file and stores it on the server
      parameters:
      - description: Event Image
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Upload a file
      tags:
      - Events
  /users/add:
    post:
      consumes:
      - application/json
      description: Adds a new User to the system
      parameters:
      - description: User Data
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      summary: Add a new User
      tags:
      - Users
  /users/delete:
    delete:
      consumes:
      - application/json
      description: Deletes a User from the system by ID
      parameters:
      - description: User ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User deleted successfully
          schema:
            type: string
      summary: Delete a User
      tags:
      - Users
  /users/get:
    get:
      consumes:
      - application/json
      description: Retrieves user details from the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
      summary: Get user details
      tags:
      - Users
swagger: "2.0"